+++ src/client/provider.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <link rel="icon" href="favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Adaptive Platform Provider</title>
+</head>
+
+<body>
+  <div id="app">
+    <h1>Adaptive Platform Provider</h1>
+    <p>
+      This window should only be visible during local development to help
+      debugging.
+    </p>
+    <p>
+      It should be running as a headless application in the environments and
+      you shouldn't be seeing this.
+    </p>
+  </div>
+  <script type="module" src="/src/OpenFin/provider.ts"></script>
+</body>
+
+</html>
