+++ src/client/public-openfin/launcher.json
@@ -28,12 +28,7 @@
     "version": "<OPENFIN_RUNTIME>",
     "arguments": "--remote-debugging-port=9091"
   },
-  "services": [
-    {
-      "name": "notifications",
-      "manifestUrl": "https://cdn.openfin.co/services/openfin/notifications/1.20.0/app.json"
-    }
-  ],
+  "services": [{ "name": "notifications" }],
   "shortcut": {
     "company": "Adaptive Financial Consulting",
     "icon": "<BASE_URL>/static/media/adaptive.ico",
@@ -45,12 +40,6 @@
       "alias": "excel-api-addin",
       "version": "2.0.0",
       "forceDownload": true
-    },
-    {
-      "src": "<BASE_URL>/plugin/LimitChecker.zip",
-      "alias": "LimitChecker",
-      "version": "1.6.0",
-      "target": "LimitChecker.application"
     }
   ],
   "snapshot": {
