+++ src/client/public-openfin/rt-fx.json
@@ -5,7 +5,7 @@
     "name": "Reactive Trader® FX <ENV_SUFFIX>",
     "icon": "<BASE_URL>/static/media/reactive-trader-icon-256x256.png",
     "url": "<BASE_URL>",
-    "autoShow": false,
+    "providerUrl": "<BASE_URL>/provider.html",
     "defaultWindowOptions": {
       "contextMenu": true,
       "fdc3InteropApi": "1.2",
@@ -22,12 +22,7 @@
     "version": "<OPENFIN_RUNTIME>",
     "arguments": "--remote-debugging-port=9091"
   },
-  "services": [
-    {
-      "name": "notifications",
-      "manifestUrl": "https://cdn.openfin.co/services/openfin/notifications/1.20.0/app.json"
-    }
-  ],
+  "services": [{ "name": "notifications" }],
   "splashScreenImage": "<BASE_URL>/static/media/splash.png",
   "shortcut": {
     "company": "Adaptive Financial Consulting",
