+++ src/client/src/App/Trades/CoreFxTrades.tsx
@@ -7,6 +7,12 @@ import { TradesGrid } from "./TradesGrid"
 import { useFxTradeRowHighlight } from "./TradesState"
 import { fxColDef, fxColFields } from "./TradesState/colConfig"
 
+if (window.fdc3) {
+  window.fdc3.addIntentListener("ViewChart", (e) => {
+    console.debug("ViewChart intent received", e)
+  })
+}
+
 const FxTrades = () => {
   const highlightedRow = useFxTradeRowHighlight()
 
@@ -17,6 +23,7 @@ const FxTrades = () => {
     }
     if (window.fdc3) {
       broadcast(context)
+      window.fdc3.raiseIntent("ViewChart", context)
     }
   }, [])
