+++ src/workspace/package.json
@@ -1,6 +1,6 @@
 {
   "name": "adaptive-workspace-provider",
-  "version": "11.0.6",
+  "version": "13.0.7",
   "description": "Adaptive Workspace Provider",
   "scripts": {
     "start": "vite",
